added real distance

+++ Android/app/src/main/java/com/app/runtogether/db/run/Run.kt
@@ -9,7 +9,7 @@ data class Run(
     @PrimaryKey(autoGenerate = true) val run_id: Int = 0,
     @ColumnInfo(name = "city") val city: String?,
     @ColumnInfo(name = "description") val description: String?,
-    @ColumnInfo(name = "length_km") val length_km: Int?,
+    @ColumnInfo(name = "length_km") val length_km: Double?,
     @ColumnInfo(name="day") val day: Long?,
     @ColumnInfo(name = "polyline") val polyline : String?
 )
